+++ settings.lua
@@ -1,0 +1,11 @@
+data:extend({
+  {
+    type = "double-setting",
+    name = "logicarts-fuel-threshold",
+    order = "a",
+    setting_type = "runtime-global",
+    default_value = 0.25,
+    minimum_value = 0.1,
+    maximum_value = 1.0,
+  },
+})